Move logic out of handlers
