+++ crates/module-system/module-implementations/sov-evm/src/rpc/handlers.rs
@@ -1,5 +1,4 @@
 use crate::db::commit::FallibleDatabaseCommit;
-use crate::error::into_rpc_error;
 use crate::rpc::error::ensure_success;
 use alloy_primitives::{Address, U64};
 use alloy_primitives::{Bytes, B256, U256};
@@ -16,12 +15,9 @@ use sov_address::{EthereumAddress, FromVmAddress};
 use sov_modules_api::macros::{config_value, rpc_gen};
 use sov_modules_api::prelude::UnwrapInfallible;
 use sov_modules_api::{ApiStateAccessor, GasMeter, GasSpec, Spec};
-use sov_rollup_interface::common::RollupHeight;
 use sov_rpc_eth_types::EthApiError;
 use tracing::debug;
 
-use crate::conversions::replay_tx_env;
-use crate::executor::{get_cfg_env, transact_commit};
 use crate::{apply_margins, Evm};
 use std::ops::DerefMut;
 
@@ -301,46 +297,7 @@ where
         state: &mut ApiStateAccessor<S>,
     ) -> RpcResult<Vec<TraceResult>> {
         debug!("EVM module JSON-RPC request to `debug_traceBlockByNumber`");
-        let opts = opts.unwrap_or_default();
-        let block_number = self.resolve_block_number(block, state);
-        // Get transaction and block data
-        let block = self
-            .blocks
-            .get(&block_number, state)?
-            .expect("Block not available");
-
-        // Get archival state and environment
-        let mut archival_state = state
-            .get_archival_state(RollupHeight::new(block_number - 1))
-            .map_err(into_rpc_error)?;
-
-        let block_env = self
-            .block_env(&mut archival_state)?
-            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-        let cfg = self.cfg(&mut archival_state).map_err(into_rpc_error)?;
-        let cfg_env = get_cfg_env(&block_env, cfg, None);
-
-        // Replay previous transactions in the block
-        let mut evm_db = self.get_db(&mut archival_state);
-
-        let mut traces = vec![];
-        for tx_idx in block.transactions {
-            let tx = self
-                .transaction(tx_idx, state)
-                .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-            let result = self.trace_transaction(
-                &block_env,
-                replay_tx_env(&tx),
-                cfg_env.clone(),
-                &mut evm_db,
-                &opts,
-            )?;
-            traces.push(TraceResult::new_success(result, Some(*tx.hash())));
-        }
-
-        Ok(traces)
+        Ok(self.trace_block_by_number(block, opts.unwrap_or_default(), state)?)
     }
 
     /// Handler for: `debug_traceTransaction`
@@ -352,56 +309,6 @@ where
         state: &mut ApiStateAccessor<S>,
     ) -> RpcResult<GethTrace> {
         debug!("EVM module JSON-RPC request to `debug_traceTransaction`");
-        // Get transaction and block data
-        let index = self
-            .get_tx_index_by_hash(&tx_hash, state)
-            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-        let traced_tx = self
-            .transaction(index, state)
-            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-        let block = self
-            .blocks
-            .get(&traced_tx.block_number, state)?
-            .expect("Transaction block not available");
-
-        // Get archival state and environment
-        let mut archival_state = state
-            .get_archival_state(RollupHeight::new(traced_tx.block_number - 1))
-            .map_err(into_rpc_error)?;
-
-        let block_env = self
-            .block_env(&mut archival_state)?
-            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-        let cfg = self.cfg(&mut archival_state).map_err(into_rpc_error)?;
-        let cfg_env = get_cfg_env(&block_env, cfg, None);
-
-        // Replay previous transactions in the block
-        let mut evm_db = self.get_db(&mut archival_state);
-
-        for tx_idx in block.transactions {
-            let tx = self
-                .transaction(tx_idx, state)
-                .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
-
-            // Skip the transaction we're tracing
-            if *tx.signed_transaction.hash() == tx_hash {
-                break;
-            }
-
-            transact_commit(&mut evm_db, &block_env, replay_tx_env(&tx), cfg_env.clone())
-                .map_err(EthApiError::from)?;
-        }
-
-        // Trace the target transaction
-        Ok(self.trace_transaction(
-            &block_env,
-            replay_tx_env(&traced_tx),
-            cfg_env,
-            &mut evm_db,
-            &opts.unwrap_or_default(),
-        )?)
+        Ok(self.trace_transaction(tx_hash, opts.unwrap_or_default(), state)?)
     }
 }
