+++ crates/module-system/module-implementations/sov-evm/src/rpc/mod.rs
@@ -8,9 +8,9 @@ use alloy_rpc_types::{
     TransactionReceipt, TransactionRequest,
 };
 use alloy_rpc_types::{BlockTransactionsKind, Header};
-use alloy_rpc_types_trace::geth::GethDebugTracingOptions;
 use alloy_rpc_types_trace::geth::GethTrace;
 use alloy_rpc_types_trace::geth::{GethDebugBuiltInTracerType, GethDebugTracerType};
+use alloy_rpc_types_trace::geth::{GethDebugTracingOptions, TraceResult};
 use jsonrpsee::core::RpcResult;
 use revm::context::result::ResultAndState;
 use revm::context::{BlockEnv, CfgEnv, TxEnv};
@@ -22,11 +22,12 @@ use sov_modules_api::{ApiStateAccessor, Spec};
 use sov_rollup_interface::common::RollupHeight;
 use sov_rpc_eth_types::{EthApiError, RpcInvalidTransactionError};
 
+use crate::conversions::replay_tx_env;
 use crate::db::EvmDb;
 use crate::error::into_rpc_error;
 use crate::evm::executor;
 use crate::evm::primitive_types::{Receipt, TransactionSigned, TxSignedAndRecovered};
-use crate::executor::{get_cfg_env, inspect};
+use crate::executor::{get_cfg_env, inspect, transact_commit};
 use crate::helpers::{from_recovered_with_block_context, prepare_call_env};
 pub use crate::primitive_types::MaybeSealedBlock;
 use crate::Evm;
@@ -173,7 +174,112 @@ where
         Some(receipts)
     }
 
+    fn trace_block_by_number(
+        &self,
+        block: BlockNumberOrTag,
+        opts: GethDebugTracingOptions,
+        state: &mut ApiStateAccessor<S>,
+    ) -> Result<Vec<TraceResult>, EthApiError> {
+        let block_number = self.resolve_block_number(block, state);
+        // Get transaction and block data
+        let block = self
+            .blocks
+            .get(&block_number, state)?
+            .ok_or(EthApiError::PrunedHistoryUnavailable)?;
+
+        // Get archival state and environment
+        let mut archival_state = state
+            .get_archival_state(RollupHeight::new(block_number - 1))
+            .map_err(|e| EthApiError::other(into_rpc_error(e)))?;
+
+        let block_env = self
+            .block_env(&mut archival_state)?
+            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+        let cfg = self.cfg(&mut archival_state)?;
+        let cfg_env = get_cfg_env(&block_env, cfg, None);
+
+        // Replay previous transactions in the block
+        let mut evm_db = self.get_db(&mut archival_state);
+
+        let mut traces = vec![];
+        for tx_idx in block.transactions {
+            let tx = self
+                .transaction(tx_idx, state)
+                .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+            let result = self.trace_transaction_inner(
+                &block_env,
+                replay_tx_env(&tx),
+                cfg_env.clone(),
+                &mut evm_db,
+                &opts,
+            )?;
+            traces.push(TraceResult::new_success(result, Some(*tx.hash())));
+        }
+        Ok(traces)
+    }
+
     fn trace_transaction(
+        &self,
+        tx_hash: B256,
+        opts: GethDebugTracingOptions,
+        state: &mut ApiStateAccessor<S>,
+    ) -> Result<GethTrace, EthApiError> {
+        // Get transaction and block data
+        let index = self
+            .get_tx_index_by_hash(&tx_hash, state)
+            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+        let traced_tx = self
+            .transaction(index, state)
+            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+        let block = self
+            .blocks
+            .get(&traced_tx.block_number, state)?
+            .expect("Transaction block not available");
+
+        // Get archival state and environment
+        let mut archival_state = state
+            .get_archival_state(RollupHeight::new(traced_tx.block_number - 1))
+            .map_err(|e| EthApiError::other(into_rpc_error(e)))?;
+
+        let block_env = self
+            .block_env(&mut archival_state)?
+            .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+        let cfg = self.cfg(&mut archival_state)?;
+        let cfg_env = get_cfg_env(&block_env, cfg, None);
+
+        // Replay previous transactions in the block
+        let mut evm_db = self.get_db(&mut archival_state);
+
+        for tx_idx in block.transactions {
+            let tx = self
+                .transaction(tx_idx, state)
+                .ok_or_else(|| EthApiError::PrunedHistoryUnavailable)?;
+
+            // Skip the transaction we're tracing
+            if *tx.signed_transaction.hash() == tx_hash {
+                break;
+            }
+
+            transact_commit(&mut evm_db, &block_env, replay_tx_env(&tx), cfg_env.clone())
+                .map_err(EthApiError::from)?;
+        }
+
+        // Trace the target transaction
+        Ok(self.trace_transaction_inner(
+            &block_env,
+            replay_tx_env(&traced_tx),
+            cfg_env,
+            &mut evm_db,
+            &opts,
+        )?)
+    }
+
+    fn trace_transaction_inner(
         &self,
         block_env: &BlockEnv,
         tx_env: TxEnv,
